requirements.txt
langchain>=0.0.310
langchain-openai>=0.0.2
langchain-community>=0.0.10
mysql-connector-python>=8.0.33
streamlit>=1.27.0
openai>=1.3.0
python-dotenv>=1.0.0
tabulate>=0.9.0
sqlalchemy>=2.0.22
matplotlib>=3.7.0
pandas>=1.5.3
plotly>=5.14.0 